move change items to init
