+++ exchangelib/folders/known_folders.py
@@ -1,6 +1,5 @@
 from ..items import CalendarItem, Contact, Message, Task, DistributionList, MeetingRequest, MeetingResponse, \
-    MeetingCancellation, CreateChange, UpdateChange, DeleteChange, ReadFlagChange, ITEM_CLASSES, ALL_ITEM_CLASSES, \
-    ASSOCIATED
+    MeetingCancellation, CreateChange, UpdateChange, DeleteChange, ReadFlagChange, ITEM_CLASSES, ASSOCIATED
 from ..version import EXCHANGE_2010_SP1, EXCHANGE_2013, EXCHANGE_2013_SP1
 from .base import Folder
 from .collections import FolderCollection
