+++ exchangelib/items/__init__.py
@@ -1,3 +1,4 @@
+from ..fields import EWSElementField
 from .base import RegisterMixIn, MESSAGE_DISPOSITION_CHOICES, SAVE_ONLY, SEND_ONLY, SEND_AND_SAVE_COPY
 from .calendar_item import CalendarItem, AcceptItem, TentativelyAcceptItem, DeclineItem, CancelCalendarItem, \
     MeetingRequest, MeetingResponse, MeetingCancellation, CONFERENCE_TYPES
@@ -6,8 +7,7 @@ from .item import SEND_MEETING_INVITATIONS_CHOICES, SEND_TO_NONE, SEND_ONLY_TO_A
     SEND_MEETING_INVITATIONS_AND_CANCELLATIONS_CHOICES, SEND_ONLY_TO_CHANGED, SEND_TO_CHANGED_AND_SAVE_COPY, \
     SEND_MEETING_CANCELLATIONS_CHOICES, AFFECTED_TASK_OCCURRENCES_CHOICES, ALL_OCCURRENCIES, \
     SPECIFIED_OCCURRENCE_ONLY, CONFLICT_RESOLUTION_CHOICES, NEVER_OVERWRITE, AUTO_RESOLVE, ALWAYS_OVERWRITE, \
-    DELETE_TYPE_CHOICES, HARD_DELETE, SOFT_DELETE, MOVE_TO_DELETED_ITEMS, BaseItem, Item, BulkCreateResult, \
-    ReadFlagChange, CreateChange, UpdateChange, DeleteChange
+    DELETE_TYPE_CHOICES, HARD_DELETE, SOFT_DELETE, MOVE_TO_DELETED_ITEMS, BaseItem, Item, BulkCreateResult
 from .message import Message, ReplyToItem, ReplyAllToItem, ForwardItem
 from .post import PostItem, PostReplyItem
 from .task import Task
@@ -25,6 +25,7 @@ __all__ = [
     'BulkCreateResult',
     'Message', 'ReplyToItem', 'ReplyAllToItem', 'ForwardItem',
     'PostItem', 'PostReplyItem',
+    'ReadFlagChange', 'CreateChange', 'UpdateChange', 'DeleteChange',
     'Task',
 ]
 
@@ -48,6 +49,40 @@ ACTIVE_DIRECTORY_CONTACTS = 'ActiveDirectoryContacts'
 CONTACTS = 'Contacts'
 CONTACTS_ACTIVE_DIRECTORY = 'ContactsActiveDirectory'
 SEARCH_SCOPE_CHOICES = (ACTIVE_DIRECTORY, ACTIVE_DIRECTORY_CONTACTS, CONTACTS, CONTACTS_ACTIVE_DIRECTORY)
+
+
+class ReadFlagChange(Item):
+    """
+    MSDN: https://msdn.microsoft.com/en-us/library/office/aa565609(v=exchg.150).aspx
+    """
+    ELEMENT_NAME = 'ReadFlagChange'
+    FIELDS = Item.FIELDS
+
+
+class CreateChange(Item):
+    """
+    MSDN: https://msdn.microsoft.com/en-us/library/office/aa565609(v=exchg.150).aspx
+    """
+    ELEMENT_NAME = 'Create'
+    FIELDS = Item.FIELDS + [
+        EWSElementField('changed_item', field_uri='Message', value_cls=Message)
+    ]
+
+
+class UpdateChange(Item):
+    """
+    MSDN: https://msdn.microsoft.com/en-us/library/office/aa565609(v=exchg.150).aspx
+    """
+    ELEMENT_NAME = 'Update'
+    FIELDS = Item.FIELDS
+
+
+class DeleteChange(Item):
+    """
+    MSDN: https://msdn.microsoft.com/en-us/library/office/aa565609(v=exchg.150).aspx
+    """
+    ELEMENT_NAME = 'Delete'
+    FIELDS = Item.FIELDS
 
 
 ITEM_CLASSES = (Item, CalendarItem, Contact, DistributionList, Message, PostItem, Task, MeetingRequest, MeetingResponse,
